Fix race condition when using multiple runners
Previously, when using multiple runners in the same process (such as
a web and worker process), there was a single mapping of files to their
mtimes, so the first runner to check a given file would recompile, but
any others wouldn't pick up the changes.

This commit moves the mtime mapping to be a property of the runner
itself so that one runner won't overwrite the mtime of a file that
another runner is also monitoring.